+++ src/RJDummy.h
@@ -11,11 +11,11 @@ void		RJDummyCopy(RATES *A, RATES *B);
 
 void		MapDummyCodes(TREES *Trees, RATES *Rates);
 
-void		RJDummyMove(int Itters, OPTIONS *Opt, TREES *Trees, RATES *Rates);
+void		RJDummyMove(long long Itters, OPTIONS *Opt, TREES *Trees, RATES *Rates);
 void		RJDummyMoveNode(OPTIONS *Opt, TREES *Trees, RATES *Rates);
 
 void		InitRJDummyFile(OPTIONS *Opt);
-void		PrintRJDummy(int Itter, OPTIONS *Opt, TREES *Trees, RATES *Rates);
+void		PrintRJDummy(long long Itter, OPTIONS *Opt, TREES *Trees, RATES *Rates);
 
 void		BuildDummyCodeBeta(RJDUMMY *RJDummy);
