+++ src/rates.h
@@ -13,7 +13,7 @@ void		CopyRates(RATES *A, RATES *B, OPTIONS *Opt);
 
 int			FindNoOfRates(OPTIONS *Opt);
 int			FindRatePos(int Rate, OPTIONS *Opt);
-void		MutateRates(OPTIONS* Opt, RATES* Rates, SCHEDULE* Shed, int It);
+void		MutateRates(OPTIONS* Opt, RATES* Rates, SCHEDULE* Shed, long long It);
 void		MutateHetero(RATES *Rates);
 void		FreeRates(RATES *Rates);
